backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymupdf==1.23.19
transformers==4.36.2
torch>=2.6.0
sentence-transformers>=2.7.0
faiss-cpu==1.11.0
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy>=1.25.0,<2.0.0
aiofiles==23.2.1
pydantic==2.5.0 